Resume-Prediction-main/Resume-Prediction-main/requirements.txt
streamlit==1.24.0
Pillow==9.2.0
pickle-mixin==1.0.2
nltk==3.8.1
